:recycle: (presenter) move present functions
